// src/app/pages/liquidacions/components/liquidacions-table/liquidacions.interface.ts
export interface LiquidacionsInterface {
   idliquidacion?: number;
   cantidadRecibida?: number;
   cambio?: number;
   folio?: string;
   kilometraje?: number;
   fecha?: string;
   nota?: string;
   cantPagada?: number;
   cantDeuda?: number;
   status?: string;
   bonificado?: number;
   descripcion?: string;
   chofer_idchofer?: number;
   baja?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}